refactor: replace type assertions with runtime validation and rename test helper
- Replace `as ScopeRole` type assertions with `scopeRoleSchema.parse()`
  in scope-member-service.ts and resolve-scope.ts for fail-fast on
  unexpected database values
- Rename `org-test-helpers.ts` to `clerk-org-mock.ts` for naming
  consistency after org-to-scope migration
