+++ turbo/apps/web/app/api/scope/__tests__/remove-member.test.ts
@@ -6,7 +6,7 @@ import { GET as getMembersRoute } from "../members/route";
 import { createTestRequest } from "../../../../src/__tests__/api-test-helpers";
 import { testContext, uniqueId } from "../../../../src/__tests__/test-helpers";
 import { mockClerk } from "../../../../src/__tests__/clerk-mock";
-import { setupClerkOrgMock } from "../../../../src/__tests__/org-test-helpers";
+import { setupClerkOrgMock } from "../../../../src/__tests__/clerk-org-mock";
 import { clerkClient } from "@clerk/nextjs/server";
 
 const context = testContext();
