+++ turbo/apps/web/src/lib/scope/scope-member-service.ts
@@ -2,7 +2,7 @@ import { eq, and, asc } from "drizzle-orm";
 import { scopeMembers } from "../../db/schema/scope-member";
 import { scopes } from "../../db/schema/scope";
 import { forbidden, notFound } from "../errors";
-import type { ScopeRole } from "@vm0/core";
+import { scopeRoleSchema, type ScopeRole } from "@vm0/core";
 
 /**
  * Get a scope member record for a specific user in a scope
@@ -28,7 +28,7 @@ export async function requireScopeMember(scopeId: string, userId: string) {
   if (!member) {
     throw forbidden("You are not a member of this scope");
   }
-  return { ...member, role: member.role as ScopeRole };
+  return { ...member, role: scopeRoleSchema.parse(member.role) };
 }
 
 /**
